@@ -60,23 +60,25 @@
     "@graphql-codegen/typescript-operations": "^3.0.2",
     "@paypal/react-paypal-js": "^8.7.0",
     "@tailwindcss/typography": "^0.5.13",
-    "@testing-library/dom": "^9.3.4",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
     "@types/node": "^18.14.6",
     "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-input-mask": "^3.0.1",
     "autoprefixer": "^10.4.2",
     "concurrently": "^8.0.0",
-    "eslint": "^8.6.0",
+    "eslint": "8.56.0",
     "eslint-config-custom": "workspace:*",
     "eslint-plugin-tailwindcss": "^3.6.2",
     "intersection-observer": "^0.12.2",
-    "jest": "^29.1.2",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.1.2",
     "jest-transform-stub": "^2.0.0",
-    "msw": "^1.0.0",
+    "msw": "1.3.5",
     "next-router-mock": "^0.9.2",
     "postcss": "^8.4.4",
     "resize-observer-polyfill": "^1.5.1",
