@@ -18,21 +18,19 @@
     "@storybook/addon-interactions": "^7.0.0-beta.38",
     "@storybook/react": "^7.0.0-beta.38",
     "@storybook/react-vite": "^7.0.0-beta.38",
-    "@testing-library/dom": "^9.0.0",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "concurrently": "^8.0.0",
-    "eslint": "^8.24.0",
+    "eslint": "8.56.0",
     "eslint-config-custom": "workspace:*",
-    "jest": "^29.1.2",
+    "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.1.2",
     "postcss": "^8.4.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "storybook": "^7.0.0-beta.38",
     "tailwindcss": "^3.4.0",
     "ts-jest": "^29.1.2",
@@ -45,5 +43,9 @@
     "csstype": "^3.0.10",
     "react-hot-toast": "^2.4.0",
     "react-icons": "^4.11.0"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
