@@ -1,10 +1,10 @@
-/* eslint-disable unicorn/filename-case */
-/* eslint-disable unicorn/prefer-module */
+import '@testing-library/jest-dom'
+import 'whatwg-fetch'
+import 'intersection-observer'
 
-require('@testing-library/jest-dom')
-require('whatwg-fetch')
-require('intersection-observer')
-global.ResizeObserver = require('resize-observer-polyfill')
+import resizeObserver from 'resize-observer-polyfill'
+
+global.ResizeObserver = resizeObserver
 
 jest.mock('next/image', () => ({
   __esModule: true,
@@ -13,6 +13,7 @@ jest.mock('next/image', () => ({
   },
 }))
 
+// eslint-disable-next-line unicorn/prefer-module
 jest.mock('next/router', () => require('next-router-mock'))
 
 Object.defineProperty(window, 'matchMedia', {
