@@ -3,11 +3,11 @@ import { useForm } from 'react-hook-form'
 import { InputField } from '@progwise/timebook-ui'
 
 import { FragmentType, graphql, useFragment } from '../../generated/gql'
-import { OrganizationInput } from '../../mocks/mocks.generated'
 import { PageHeading } from '../pageHeading'
 import { ArchiveOrUnarchiveOrganizationButton } from './archiveOrUnarchiveOrganizationButton/archiveOrUnarchiveOrganizationButton'
 import { OrganizationSubscriptionStatusLabel } from './organizationSubscriptionStatusLabel'
 import { SubscribeOrUnsubscribeOrganizationButton } from './subscribeOrUnsubscribeOrganizationButton/subscribeOrUnsubscribeOrganizationButton'
+import { OrganizationInput } from '../../generated/gql/graphql'
 
 export const OrganizationFormFragment = graphql(`
   fragment OrganizationForm on Organization {
