@@ -7,7 +7,7 @@ import { makeFragmentData } from '../../generated/gql'
 import '../../mocks/mockServer'
 import { TaskRow, TaskRowFragment } from './taskRow'
 
-const client = new Client({ url: '/api/graphql' })
+const client = new Client({ url: '/api/graphql', exchanges: [] })
 const wrapper: React.FC<{ children: React.ReactNode }> = ({ children }) => (
   <Provider value={client}>{children}</Provider>
 )
